mic05 project/backend/requirements.txt
# ==================================
# 核心Web框架
# ==================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# ==================================
# 数据库相关 (Tortoise ORM)
# ==================================
tortoise-orm==0.20.0
asyncmy==0.2.9
aerich==0.7.2

# ==================================
# AI服务集成
# ==================================
openai==1.3.7
httpx==0.25.2
aiohttp==3.9.1

# ==================================
# 音频处理
# ==================================
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4

# ==================================
# 数据验证和序列化
# ==================================
pydantic==2.5.0
python-multipart==0.0.6

# ==================================
# 文本和语言处理
# ==================================
langdetect==1.0.9
python-magic==0.4.27

# ==================================
# 文件处理和导出
# ==================================
markdown==3.5.1
python-docx==1.1.0
openpyxl==3.1.2

# ==================================
# 缓存和存储
# ==================================
aioredis==2.0.1
aiofiles==23.2.0

# ==================================
# 工具库
# ==================================
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3

# ==================================
# 日志和监控
# ==================================
loguru==0.7.2
prometheus-client==0.19.0

# ==================================
# 开发和测试
# ==================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# ==================================
# 云服务SDK（可选，根据实际使用的服务选择）
# ==================================
# Azure认知服务
azure-cognitiveservices-speech==1.34.0
azure-ai-translation-text==1.0.0

# Google Cloud服务
google-cloud-speech==2.21.0
google-cloud-translate==3.12.1
google-cloud-texttospeech==2.16.3

# AWS服务
boto3==1.34.0
botocore==1.34.0

# ==================================
# 安全和认证
# ==================================
cryptography==41.0.7
bcrypt==4.1.2

# ==================================
# 环境和配置
# ==================================
pyyaml==6.0.1
toml==0.10.2